pyproject.toml
[project]
name = "ai-employee"
version = "0.1.0"
description = "Personal AI Employee - Autonomous Digital FTE using Claude Code + Obsidian"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "AI Employee Project" }
]
keywords = ["ai", "automation", "claude", "obsidian", "mcp"]

dependencies = [
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "schedule>=1.2.0",
]

[project.optional-dependencies]
google = [
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
]
browser = [
    "playwright>=1.40.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
all = [
    "ai-employee[google,browser,dev]",
]

[project.scripts]
ai-employee = "src.orchestrator:main"
fs-watcher = "src.watchers.filesystem_watcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
